# OAK-11154: Read partial segments from SegmentWriter