--- a/oak-segment-tar/src/test/java/org/apache/jackrabbit/oak/segment/SegmentBufferWriterTest.java
+++ b/oak-segment-tar/src/test/java/org/apache/jackrabbit/oak/segment/SegmentBufferWriterTest.java
@@ -20,8 +20,11 @@ package org.apache.jackrabbit.oak.segment;
 import static org.apache.jackrabbit.guava.common.collect.Lists.newArrayList;
 import static org.apache.jackrabbit.oak.segment.DefaultSegmentWriterBuilder.defaultSegmentWriterBuilder;
 import static org.apache.jackrabbit.oak.segment.file.FileStoreBuilder.fileStoreBuilder;
+import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 
 import java.io.File;
 import java.util.Collections;
@@ -29,8 +32,10 @@ import java.util.List;
 import java.util.Optional;
 
 import org.apache.jackrabbit.oak.plugins.memory.EmptyNodeState;
+import org.apache.jackrabbit.oak.segment.data.PartialSegmentState;
 import org.apache.jackrabbit.oak.segment.file.FileStore;
 import org.apache.jackrabbit.oak.segment.file.ReadOnlyFileStore;
+import org.apache.jackrabbit.oak.segment.file.tar.GCGeneration;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -135,4 +140,74 @@ public class SegmentBufferWriterTest {
         }
     }
 
+    @Test
+    public void readPartialSegmentState() throws Exception {
+        try (FileStore store = openFileStore()) {
+            GCGeneration gcGeneration = store.getRevisions().getHead().getSegment().getGcGeneration();
+            SegmentBufferWriter writer = new SegmentBufferWriter(store.getSegmentIdProvider(), "t", gcGeneration);
+
+            var referencedRecordId = store.getHead().getRecordId();
+            writer.flush(store);
+
+            // Write a record completely
+            RecordId completeRecordId = writer.prepare(RecordType.BLOCK, Long.BYTES, Collections.emptyList(), store);
+            writer.writeLong(0b00001010_11010000_10111110_10101101_00001011_11101010_11010000_10111110L);
+            assert completeRecordId.getSegmentId() != referencedRecordId.getSegmentId();
+
+            // Reference another segment
+            final int RECORD_ID_BYTES = Short.BYTES + Integer.BYTES;
+            RecordId referenceRecordId = writer.prepare(RecordType.VALUE, RECORD_ID_BYTES, Collections.emptyList(), store);
+            writer.writeRecordId(referencedRecordId);
+
+            // Write a record partially
+            RecordId partialRecordId = writer.prepare(RecordType.BLOCK, Integer.BYTES * 2, Collections.emptyList(), store);
+            writer.writeByte((byte)42);
+
+            // Read the segment partially
+            PartialSegmentState partialSegmentState = writer.readPartialSegmentState(completeRecordId.getSegmentId());
+            assertNotNull(partialSegmentState);
+
+            assertEquals(partialSegmentState.generation(), gcGeneration.getGeneration());
+            assertEquals(partialSegmentState.fullGeneration(), gcGeneration.getFullGeneration());
+            assertEquals(partialSegmentState.isCompacted(), gcGeneration.isCompacted());
+            assertEquals(partialSegmentState.version(), (byte) 13);
+
+            assertEquals(1, partialSegmentState.segmentReferences().size());
+            assertEquals(referencedRecordId.getSegmentId().getMostSignificantBits(), partialSegmentState.segmentReferences().get(0).msb());
+            assertEquals(referencedRecordId.getSegmentId().getLeastSignificantBits(), partialSegmentState.segmentReferences().get(0).lsb());
+
+            assertEquals(4, partialSegmentState.records().size());
+
+            writer.flush(store);
+
+            int offset = partialSegmentState.records().get(1).offset();
+            assertEquals(completeRecordId.getRecordNumber(), partialSegmentState.records().get(1).refNumber());
+            assertEquals(RecordType.BLOCK, partialSegmentState.records().get(1).recordType());
+            assertArrayEquals(new byte[] { (byte)0b00001010, (byte)0b11010000, (byte)0b10111110, (byte)0b10101101, (byte)0b00001011, (byte)0b11101010, (byte)0b11010000, (byte)0b10111110 }, partialSegmentState.records().get(1).contents());
+            offset -= Long.BYTES;
+
+            assertEquals(referenceRecordId.getRecordNumber(), partialSegmentState.records().get(2).refNumber());
+            assertEquals(RecordType.VALUE, partialSegmentState.records().get(2).recordType());
+            assertEquals(offset, partialSegmentState.records().get(2).offset());
+            offset -= 8; // align(RECORD_ID_BYTES, 4)
+
+            assertEquals(partialRecordId.getRecordNumber(), partialSegmentState.records().get(3).refNumber());
+            assertEquals(RecordType.BLOCK, partialSegmentState.records().get(3).recordType());
+            assertEquals(offset, partialSegmentState.records().get(3).offset());
+            assertArrayEquals(new byte[] { 42 }, partialSegmentState.records().get(3).contents());
+        }
+    }
+
+    @Test
+    public void readPartialSegmentStateOfOtherSegmentReturnsNull() throws Exception {
+        try (FileStore store = openFileStore()) {
+            SegmentBufferWriter writer = new SegmentBufferWriter(store.getSegmentIdProvider(), "t", store.getRevisions().getHead().getSegment().getGcGeneration());
+
+            RecordId recordId = writer.prepare(RecordType.BLOCK, Long.BYTES, Collections.emptyList(), store);
+            writer.writeLong(0b101011010000101111101010110100001011111010101101000010111110L);
+            writer.flush(store);
+
+            assertNull(writer.readPartialSegmentState(recordId.getSegmentId()));
+        }
+    }
 }